Simonas_meal_planner/requirements.txt
pytest==7.2.1
pytest-mock==3.10.0
requests==2.31.0
python-dotenv==1.0.0
